requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
gunicorn==20.1.0
openai==0.27.0
pyngrok==7.2.1
python-dotenv
streamlit